Duke, Descent 'II errors scrubbed!
Dealt with the slowdown issues on Atomic Edition and updated it to prevent crashing! Normally it crashes after an hour of gameplay, well no more!

Likewise I swapped the CFG for Descent1 to fix the intro from Descent2. Before it would crash if you let the intro play out, these newer CFGs should take care of that.

This is my first github fix so, please give feedback and enjoy,

I left the folder structure the same as the originals. Just dump and overwrite to the "games" folder in Flynns Top 300 pack. Profit!